# src/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ArtHub API",
    "description": "Complete API documentation for ArtHub backend services",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Main API server"
    }
  ],
...
  "paths": {
...
    "/api/chat": {
      "get": {
        "tags": [
          "Chat"
        ],
        "summary": "الحصول على محادثات المستخدم",
        "description": "استرجاع جميع محادثات المستخدم المسجل الدخول",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "تم جلب المحادثات بنجاح",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Chat"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/chat/create": {
      "post": {
        "tags": [
          "Chat"
        ],
        "summary": "إنشاء محادثة",
        "description": "إنشاء محادثة جديدة مع مستخدم آخر",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "userId"
                ],
                "properties": {
                  "userId": {
                    "type": "string",
                    "description": "معرف المستخدم الآخر"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "تم إنشاء المحادثة بنجاح",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Chat"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "المستخدم غير موجود"
          }
        }
      }
    },
    "/api/chat/{chatId}/messages": {
      "get": {
        "tags": [
          "Chat"
        ],
        "summary": "الحصول على رسائل محادثة",
        "description": "استرجاع جميع الرسائل في محادثة معينة",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "chatId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "معرف المحادثة"
          }
        ],
        "responses": {
          "200": {
            "description": "تم جلب الرسائل بنجاح",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ChatMessage"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "المحادثة غير موجودة"
          }
        }
      },
      "post": {
        "tags": [
          "Chat"
        ],
        "summary": "إرسال رسالة",
        "description": "إرسال رسالة جديدة في محادثة معينة",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "chatId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "معرف المحادثة"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "text"
                ],
                "properties": {
                  "text": {
                    "type": "string",
                    "description": "نص الرسالة"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "تم إرسال الرسالة بنجاح",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ChatMessage"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "المحادثة غير موجودة"
          }
        }
      }
    },
    "/api/chat/{chatId}/read": {
      "post": {
        "tags": [
          "Chat"
        ],
        "summary": "تحديث حالة قراءة الرسائل",
        "description": "تعيين جميع الرسائل في محادثة معينة كمقروءة",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "chatId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "معرف المحادثة"
          }
        ],
        "responses": {
          "200": {
            "description": "تم تحديث حالة القراءة بنجاح",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "modifiedCount": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "المحادثة غير موجودة"
          }
        }
      }
    },
...
    "/chat": {
      "get": {
        "tags": [
          "Chat"
        ],
        "summary": "Get user chats",
        "description": "Get all chats for the authenticated user",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of user chats",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Chat"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/chat/{chatId}/messages": {
      "get": {
        "tags": [
          "Chat"
        ],
        "summary": "Get messages in a chat",
        "description": "Get all messages in a specific chat",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "chatId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of messages",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Message"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Chat not found"
          }
        }
      }
    },
...
    "/api/follow/follow": {
      "post": {
        "tags": [
          "Follow"
        ],
        "summary": "متابعة فنان",
        "description": "إنشاء متابعة جديدة لفنان",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "artistId"
                ],
                "properties": {
                  "artistId": {
                    "type": "string",
                    "description": "معرف الفنان المراد متابعته"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "تمت المتابعة بنجاح",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowResponse"
                }
              }
            }
          },
          "400": {
            "description": "خطأ في البيانات المدخلة أو المستخدم يتابع الفنان بالفعل"
          },
          "404": {
            "description": "الفنان غير موجود"
          }
        }
      }
    },
    "/api/follow/unfollow": {
      "post": {
        "tags": [
          "Follow"
        ],
        "summary": "إلغاء متابعة فنان",
        "description": "إلغاء متابعة فنان تتم متابعته حاليًا",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "artistId"
                ],
                "properties": {
                  "artistId": {
                    "type": "string",
                    "description": "معرف الفنان المراد إلغاء متابعته"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "تم إلغاء المتابعة بنجاح",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowResponse"
                }
              }
            }
          },
          "400": {
            "description": "خطأ في البيانات المدخلة أو المستخدم لا يتابع الفنان حاليًا"
          },
          "404": {
            "description": "الفنان غير موجود"
          }
        }
      }
    },
    "/api/follow/followers/{artistId}": {
      "get": {
        "tags": [
          "Follow"
        ],
        "summary": "الحصول على متابعي فنان",
        "description": "جلب قائمة المستخدمين الذين يتابعون فنانًا معينًا",
        "parameters": [
          {
            "name": "artistId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "معرف الفنان"
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "رقم الصفحة"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            },
            "description": "عدد العناصر في الصفحة"
          }
        ],
        "responses": {
          "200": {
            "description": "تم جلب المتابعين بنجاح",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowersResponse"
                }
              }
            }
          },
          "404": {
            "description": "الفنان غير موجود"
          }
        }
      }
    },
    "/api/follow/following/{userId}": {
      "get": {
        "tags": [
          "Follow"
        ],
        "summary": "الحصول على متابَعي مستخدم",
        "description": "جلب قائمة الفنانين الذين يتابعهم المستخدم",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "معرف المستخدم"
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "رقم الصفحة"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            },
            "description": "عدد العناصر في الصفحة"
          }
        ],
        "responses": {
          "200": {
            "description": "تم جلب المتابَعين بنجاح",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowingResponse"
                }
              }
            }
          },
          "404": {
            "description": "المستخدم غير موجود"
          }
        }
      }
    },
    "/api/follow/status/{artistId}": {
      "get": {
        "tags": [
          "Follow"
        ],
        "summary": "التحقق من حالة المتابعة",
        "description": "التحقق مما إذا كان المستخدم الحالي يتابع فنانًا معينًا",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "artistId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "معرف الفنان"
          }
        ],
        "responses": {
          "200": {
            "description": "تم التحقق من حالة المتابعة بنجاح",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowStatusResponse"
                }
              }
            }
          },
          "404": {
            "description": "الفنان غير موجود"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ChatMessage": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "chatId": {
            "type": "string"
          },
          "sender": {
            "$ref": "#/components/schemas/User"
          },
          "text": {
            "type": "string"
          },
          "read": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "job": {
            "type": "string"
          },
          "profileImage": {
            "type": "object",
            "properties": {
              "url": {
                "type": "string",
                "format": "uri"
              }
            }
          }
        }
      },
...
      "Chat": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "participants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User"
            }
          },
          "lastMessage": {
            "$ref": "#/components/schemas/Message"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Message": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "chat": {
            "type": "string"
          },
          "sender": {
            "$ref": "#/components/schemas/User"
          },
          "content": {
            "type": "string"
          },
          "read": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "FollowResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "FollowersResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "object",
            "properties": {
              "followers": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      },
      "FollowingResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "object",
            "properties": {
              "following": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      },
      "FollowStatusResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "object",
            "properties": {
              "isFollowing": {
                "type": "boolean"
              }
            }
          }
        }
      },
...
    }
  }
}